Lab2 Code
To see whether h_ss predicts T(t) reasonably well, we can fit lumped-capacity model to the measurements and calculate the best-fit value of h. The curve fitting procedure is provided in the code. The resulting best-fit value of h is shown in the plot. The plot will show how well the lumped-capacity model predicts the temperature rise of the block during heating.